Improved Dynamic threshold scheme
- It turns out we do not need to simulate random noise for every transition. Something like Common Random Numbers (CRN) exist, based on which, we use same noise for all transitions belonging to same parent path.
- This improves the dynamic scheme by an order of magnitude which allows us to increase n_trials, thereby improving the accuracy.
- Adding this as "improved" mode while retaining the current mode as legacy. Hopefully, I would not have to retract it in the future.
- Added example notebook explaining threshold optimization.

+++ src/pyloki/utils/np_utils.py
@@ -95,12 +95,12 @@ def np_apply_along_axis(func1d: Callable, axis: int, arr: np.ndarray) -> np.ndar
         raise ValueError(msg)
     if axis == 0:
         res_len = arr.shape[1]
-        result = np.empty(res_len)
+        result = np.empty(res_len, dtype=arr.dtype)
         for ii in range(res_len):
             result[ii] = func1d(arr[:, ii])
     else:
         res_len = arr.shape[0]
-        result = np.empty(res_len)
+        result = np.empty(res_len, dtype=arr.dtype)
         for jj in range(res_len):
             result[jj] = func1d(arr[jj, :])
     return result
